+++ Changes.md
@@ -18,6 +18,8 @@
 ```ruby
 Sidekiq.default_worker_options = { 'queue' => 'default', 'backtrace' => true }
 ```
+- Added two handy Sidekiq::Client class methods for scheduled jobs:
+  `enqueue_to_in` and `enqueue_in` [#1212]
 
 2.14.1
 -----------