+++ Changes.md
@@ -1,3 +1,7 @@
+2.16.0
+-----------
+- Upgraded Bootstrap to version 3.0.0
+
 2.15.0
 -----------